requirements.txt
cuda-python==11.8.3
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.3.0+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.3.0+cu118

accelerate>=0.33.0
cached_path
click
datasets
einops>=0.8.0
einx>=0.3.0
ema_pytorch>=0.5.2
gradio
jieba
librosa
matplotlib
numpy<=1.26.4
pydub
pypinyin
safetensors
soundfile
tomli
torchdiffeq
tqdm>=4.65.0
transformers
vocos
wandb
x_transformers>=1.31.14

protobuf
grpcio
grpcio-tools
